# backend/src/main/java/com/pji/triagem/dto/request/UpdateUserProfileRequest.java
package com.pji.triagem.dto.request;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.Size;
import lombok.Data;

@Data
public class UpdateUserProfileRequest {

    @Size(min = 2, max = 100, message = "Nome deve ter entre 2 e 100 caracteres")
    private String name;

    @Email(message = "Email inválido")
    @Size(max = 120, message = "Email deve ter até 120 caracteres")
    private String email;
}