# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PediTriagem API
  version: 1.0.0
  description: Contrato atual do backend Spring Boot integrado ao app Expo.
servers:
  - url: http://localhost:8080
    description: Desenvolvimento local
security:
  - bearerAuth: []
tags:
  - name: Auth
  - name: Crianças
  - name: Sintomas
  - name: Avaliações
  - name: Usuários
  - name: Health
paths:
  /health:
    get:
      tags: [Health]
      summary: Verifica disponibilidade da aplicação
      security: []
      responses:
        '200':
          description: Serviço operacional
  /auth/register/user:
    post:
      tags: [Auth]
      summary: Cria usuário cuidador
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Usuário criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StringEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/login:
    post:
      tags: [Auth]
      summary: Autentica usuário por CPF/CNPJ e senha
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Tokens JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags: [Auth]
      summary: Retorna o usuário autenticado
      responses:
        '200':
          description: Perfil autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/usuarios/me:
    patch:
      tags: [Usuários]
      summary: Atualiza nome e email do usuário autenticado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserProfileRequest'
      responses:
        '200':
          description: Perfil atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /children/user/{userId}:
    get:
      tags: [Crianças]
      summary: Lista crianças de um usuário
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Crianças cadastradas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChildHomeListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /children/{userId}:
    post:
      tags: [Crianças]
      summary: Cadastra criança para um usuário
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChildRequest'
      responses:
        '201':
          description: Criança cadastrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChildEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /children/{userId}/children/{childId}:
    get:
      tags: [Crianças]
      summary: Busca criança por ID
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/ChildId'
      responses:
        '200':
          description: Criança encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChildEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
  /symptoms:
    get:
      tags: [Sintomas]
      summary: Lista sintomas disponíveis
      responses:
        '200':
          description: Catálogo de sintomas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SymptomListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /assessments/questionnaire:
    post:
      tags: [Avaliações]
      summary: Monta questionário para um ou mais sintomas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionnaireRequest'
      responses:
        '200':
          description: Questionário agrupado por sintoma
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionnaireEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
  /assessments:
    post:
      tags: [Avaliações]
      summary: Registra avaliação e calcula classificação
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssessmentRequest'
      responses:
        '201':
          description: Resultado calculado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentResultEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
  /assessments/{assessmentId}/orientation:
    get:
      tags: [Avaliações]
      summary: Retorna orientações de uma avaliação
      parameters:
        - $ref: '#/components/parameters/AssessmentId'
      responses:
        '200':
          description: Orientações agrupadas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrientationEnvelope'
        '404':
          $ref: '#/components/responses/NotFound'
  /assessments/users/{userId}/history:
    get:
      tags: [Avaliações]
      summary: Retorna histórico de avaliações por usuário
      parameters:
        - $ref: '#/components/parameters/UserId'
        - name: childId
          in: query
          required: false
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Histórico do usuário
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentHistoryEnvelope'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    ChildId:
      name: childId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    AssessmentId:
      name: assessmentId
      in: path
      required: true
      schema:
        type: integer
        format: int64
  responses:
    BadRequest:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Token ausente, inválido ou expirado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    ResponseEnvelope:
      type: object
      properties:
        data:
          nullable: true
        errors:
          type: array
          items:
            type: string
        links:
          type: array
          items:
            type: string
    ErrorEnvelope:
      allOf:
        - $ref: '#/components/schemas/ResponseEnvelope'
      example:
        data: null
        errors: ['Usuário ou senha inválidos']
        links: []
    StringEnvelope:
      allOf:
        - $ref: '#/components/schemas/ResponseEnvelope'
        - type: object
          properties:
            data:
              type: string
              example: Usuario criado com sucesso
    RegisterRequest:
      type: object
      required: [login, password, email, name]
      properties:
        login:
          type: string
          description: CPF/CNPJ sem pontuação
          example: '52998224725'
        password:
          type: string
          example: peditriagem123
        email:
          type: string
          format: email
          example: camila.demo@peditriagem.app
        name:
          type: string
          example: Camila Ribeiro
    LoginRequest:
      type: object
      required: [login, password]
      properties:
        login:
          type: string
          example: '52998224725'
        password:
          type: string
          example: peditriagem123
    TokenEnvelope:
      allOf:
        - $ref: '#/components/schemas/ResponseEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/TokenResponse'
    TokenResponse:
      type: object
      required: [accessToken, refreshToken, accessTokenExpiration, refreshTokenExpiration]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        accessTokenExpiration:
          type: integer
          format: int64
        refreshTokenExpiration:
          type: integer
          format: int64
    UserProfileEnvelope:
      allOf:
        - $ref: '#/components/schemas/ResponseEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/UserProfileResponse'
    UserProfileResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        login:
          type: string
          description: CPF/CNPJ sem pontuação
        email:
          type: string
          format: email
        name:
          type: string
    UpdateUserProfileRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: Camila Ribeiro
        email:
          type: string
          format: email
          maxLength: 120
          example: camila.ribeiro@example.com
    CreateChildRequest:
      type: object
      required: [name, birthDate]
      properties:
        name:
          type: string
          example: Maria
        cpf:
          type: string
          example: ''
        birthDate:
          type: string
          format: date
          example: '2022-05-04'
        weightKg:
          type: number
          format: double
          example: 17
        avatarEmoji:
          type: string
          example: MA
    ChildHomeListEnvelope:
      allOf:
        - $ref: '#/components/schemas/ResponseEnvelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/ChildHomeResponse'
    ChildEnvelope:
      allOf:
        - $ref: '#/components/schemas/ResponseEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/ChildResponse'
    ChildHomeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        age:
          type: string
          example: 4 anos
        ageInMonths:
          type: integer
        avatarEmoji:
          type: string
        totalAssessments:
          type: integer
          format: int64
        lastAssessmentAt:
          type: string
          format: date-time
    ChildResponse:
      allOf:
        - $ref: '#/components/schemas/ChildHomeResponse'
        - type: object
          properties:
            cpf:
              type: string
            birthDate:
              type: string
              format: date
            weightKg:
              type: number
            createdAt:
              type: string
              format: date-time
    SymptomListEnvelope:
      allOf:
        - $ref: '#/components/schemas/ResponseEnvelope'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/SymptomResponse'
    SymptomResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
          example: FEBRE
        name:
          type: string
          example: Febre
        shortDescription:
          type: string
        iconRef:
          type: string
          example: thermo
        colorHex:
          type: string
          example: '#F59E5C'
        order:
          type: integer
    QuestionnaireRequest:
      type: object
      required: [symptomIds]
      properties:
        symptomIds:
          type: array
          items:
            type: integer
            format: int64
    QuestionnaireEnvelope:
      allOf:
        - $ref: '#/components/schemas/ResponseEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/QuestionnaireResponse'
    QuestionnaireResponse:
      type: object
      properties:
        symptoms:
          type: array
          items:
            $ref: '#/components/schemas/QuestionnaireSymptomResponse'
    QuestionnaireSymptomResponse:
      type: object
      properties:
        symptomId:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
        questions:
          type: array
          items:
            $ref: '#/components/schemas/QuestionResponse'
    QuestionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        text:
          type: string
        subtitle:
          type: string
        type:
          type: string
          enum: [YESNO, OPTIONS]
        order:
          type: integer
        options:
          type: array
          items:
            $ref: '#/components/schemas/QuestionOptionResponse'
    QuestionOptionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        text:
          type: string
        order:
          type: integer
    CreateAssessmentRequest:
      type: object
      required: [childId, symptoms]
      properties:
        childId:
          type: integer
          format: int64
        symptoms:
          type: array
          items:
            $ref: '#/components/schemas/CreateAssessmentSymptomRequest'
    CreateAssessmentSymptomRequest:
      type: object
      required: [symptomId, answers]
      properties:
        symptomId:
          type: integer
          format: int64
        answers:
          type: array
          items:
            $ref: '#/components/schemas/CreateAssessmentAnswerRequest'
    CreateAssessmentAnswerRequest:
      type: object
      properties:
        questionId:
          type: integer
          format: int64
        optionId:
          type: integer
          format: int64
        answer:
          type: string
          enum: [YES, NO, DUNNO]
    AssessmentResultEnvelope:
      allOf:
        - $ref: '#/components/schemas/ResponseEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AssessmentResultResponse'
    AssessmentResultResponse:
      type: object
      properties:
        assessmentId:
          type: integer
          format: int64
        childId:
          type: integer
          format: int64
        finalClassification:
          type: string
          enum: [LOW, MOD, HIGH]
        totalScore:
          type: integer
        redFlagDetected:
          type: boolean
        symptoms:
          type: array
          items:
            $ref: '#/components/schemas/AssessmentSymptomResultResponse'
    AssessmentSymptomResultResponse:
      type: object
      properties:
        symptomId:
          type: integer
          format: int64
        symptomName:
          type: string
        classification:
          type: string
          enum: [LOW, MOD, HIGH]
        score:
          type: integer
        redFlagDetected:
          type: boolean
    OrientationEnvelope:
      allOf:
        - $ref: '#/components/schemas/ResponseEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/OrientationResponse'
    OrientationResponse:
      type: object
      properties:
        assessmentId:
          type: integer
          format: int64
        childId:
          type: integer
          format: int64
        childName:
          type: string
        finalClassification:
          type: string
          enum: [LOW, MOD, HIGH]
        title:
          type: string
        description:
          type: string
        mainActions:
          type: array
          items:
            $ref: '#/components/schemas/OrientationItemResponse'
        warningSigns:
          type: array
          items:
            $ref: '#/components/schemas/OrientationItemResponse'
        homeCare:
          type: array
          items:
            $ref: '#/components/schemas/OrientationItemResponse'
        whenSeekHelp:
          type: array
          items:
            $ref: '#/components/schemas/OrientationItemResponse'
        disclaimer:
          type: string
    OrientationItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        type:
          type: string
        symptomId:
          type: integer
          format: int64
        symptomName:
          type: string
    AssessmentHistoryEnvelope:
      allOf:
        - $ref: '#/components/schemas/ResponseEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AssessmentHistoryResponse'
    AssessmentHistoryResponse:
      type: object
      properties:
        totalAssessments:
          type: integer
          format: int64
        lowRiskCount:
          type: integer
          format: int64
        moderateRiskCount:
          type: integer
          format: int64
        highRiskCount:
          type: integer
          format: int64
        selectedChildId:
          type: integer
          format: int64
        children:
          type: array
          items:
            $ref: '#/components/schemas/AssessmentHistoryChildResponse'
        assessments:
          type: array
          items:
            $ref: '#/components/schemas/AssessmentHistoryItemResponse'
    AssessmentHistoryChildResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    AssessmentHistoryItemResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        childId:
          type: integer
          format: int64
        childName:
          type: string
        symptoms:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        classification:
          type: string
          enum: [LOW, MOD, HIGH]